refactor(settings): split Settings screen into per-section files, add Pacing section, collapse Readaloud and WebDAV behind full-screen drill-ins
Reorganises the global Settings surface into one file per section under
`sections/` and one file per drill-in panel under `panels/`. SettingsScreen.kt
shrinks from 1320 to 261 lines and becomes the scaffold that composes the
section list; shared row primitives (SettingsSectionHeader,
SettingsDrillInRow, disabledListItemColors, SettingsPanel state,
StorytellerBadge, AnnotationSyncBadge) move to SettingsCommon.kt so their
copy-paste count drops from 8/6/3 to 1.

User-visible changes:
- Auto-Scroll and Cadence move out of Reading into a new Pacing section
  (same shape, different concern: wpm-driven reading aids vs. book
  presentation).
- Readaloud and Annotations Sync collapse to a single row each. Each row
  keeps the pre-collapse badge and tone-coloured subtitle
  (Local/Synced/Pending/Error, error red, pending amber) so status stays
  legible without a tap. Tap opens a dedicated drill-in screen.

New destinations:
- settings/readaloud → ReadaloudSettingsScreen (Server / Matches / Appearance).
- settings/annotation_sync → AnnotationsSyncSettingsScreen (Server / Devices).

Regression coverage added: ReadaloudRowSummaryTest pins the collapsed row's
subtitle across not-configured, host+version, unmatched+matched, and
all-matched-collapse cases.
