+++ app/src/main/kotlin/com/riffle/app/feature/settings/SettingsNavEvent.kt
@@ -2,5 +2,4 @@ package com.riffle.app.feature.settings
 
 sealed class SettingsNavEvent {
     data object NavigateToAddSource : SettingsNavEvent()
-    data class NavigateToReadaloudMatches(val sourceId: String) : SettingsNavEvent()
 }
