+++ buildSrc/src/main/kotlin/com/riffle/buildlogic/ServerReferenceLint.kt
@@ -64,6 +64,11 @@ object ServerReferenceLint {
         "app/src/main/kotlin/com/riffle/app/feature/reader/session/PositionOrchestrator.kt",
         "app/src/main/kotlin/com/riffle/app/feature/settings/SettingsViewModel.kt",
         "app/src/main/kotlin/com/riffle/app/feature/settings/SettingsScreen.kt",
+        // Split-out sections/drill-in for Settings (post-#XXX). Same ServerType/serverType usage
+        // as SettingsScreen — routing per-source rendering and identifying the Storyteller Service.
+        "app/src/main/kotlin/com/riffle/app/feature/settings/sections/SourcesSection.kt",
+        "app/src/main/kotlin/com/riffle/app/feature/settings/sections/ReadaloudSection.kt",
+        "app/src/main/kotlin/com/riffle/app/feature/settings/readaloud/ReadaloudSettingsScreen.kt",
         "app/src/main/kotlin/com/riffle/app/feature/settings/readaloud/ReadaloudMatchesViewModel.kt",
         "app/src/main/kotlin/com/riffle/app/feature/audiobook/AudiobookPlayerViewModel.kt",
         "app/src/main/kotlin/com/riffle/app/feature/navigation/NavigationDrawerViewModel.kt",
